% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_load.R
\name{etl_load.etl_idswb}
\alias{etl_load.etl_idswb}
\title{Load the transformed data into the database.}
\usage{
\method{etl_load}{etl_idswb}(obj, ...)
}
\arguments{
\item{obj}{An etl object}

\item{...}{arguments passed to methods}
}
\description{
Load the transformed data into the database.
}
