% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_transform.R
\name{etl_transform.etl_idswb}
\alias{etl_transform.etl_idswb}
\title{Prepares raw data to be loaded into a database}
\usage{
\method{etl_transform}{etl_idswb}(obj, ...)
}
\arguments{
\item{obj}{An etl object}

\item{...}{arguments passed to methods}
}
\value{
Populates load directory with data
}
\description{
Prepares raw data to be loaded into a database

Manipulate the raw data such that it can be loaded into a database table. Usually
this means converting the raw data to (a series of) CSV files, which are also stored locally
}
