% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_debt.R
\name{get_debt}
\alias{get_debt}
\title{Gets debt information for a given country that it owes to both China and the World}
\usage{
get_debt(debtor, by_year = 2022)
}
\arguments{
\item{debtor}{Country of interest}

\item{by_year}{Year of interest}
}
\value{
Returns a tibble showing how debtor owes to both China and the World
}
\description{
This function extract a tibble showing how a debtor owes to both China and the World
}
