% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_query_data.R
\name{ids_query_data}
\alias{ids_query_data}
\title{ids_query_data
Gets debt information for a given country (debtor) that it owes to any creditor.
The default creditors are both China and the World}
\usage{
ids_query_data(etl_object, debtor, creditor = c("China", "World"), series)
}
\arguments{
\item{etl_object}{An \code{etl} object}

\item{debtor}{Debtor name(s)}

\item{creditor}{Creditor name(s)}

\item{series}{A vector of Series}
}
\value{
Returns a tibble showing how debtor owes to any creditor. The default creditors are both China and the World
}
\description{
ids_query_data
Gets debt information for a given country (debtor) that it owes to any creditor.
The default creditors are both China and the World
}
