% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R, R/etl_extract.R, R/etl_load.R,
%   R/etl_transform.R
\name{etl_extract.etl_idswb}
\alias{etl_extract.etl_idswb}
\alias{etl_load.etl_idswb}
\alias{etl_transform.etl_idswb}
\title{My ETL functions}
\usage{
\method{etl_extract}{etl_idswb}(obj, ...)

\method{etl_extract}{etl_idswb}(obj, ...)

\method{etl_load}{etl_idswb}(
  obj,
  years = lubridate::year(Sys.Date()),
  months = lubridate::month(Sys.Date()),
  ...
)

\method{etl_transform}{etl_idswb}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link[etl]{etl}} object}

\item{...}{arguments passed to methods}
}
\description{
My ETL functions

etl_extract

etl_load

etl_transform
}
\details{
This function downloads International Debt statistics data
in the raw directory under the folder the user created

This function loads idswb data into a local database for years and months specified.

This function transforms NYC311 data for years and months specified.
}
\examples{
\dontrun{
if (require(dplyr)) {
  obj <- etl("idswb") \%>\%
    etl_create()
}
}
\dontrun{
calls <- etl("idswb")
calls \%>\%
 etl_extract(years = 2010:2011, months = 1:3, num_calls = 100)

calls \%>\%
  etl_init() \%>\%
  etl_update(years = 2010:2011, months = 1:3, num_calls = 100)

calls \%>\%
  tbl("calls") \%>\%
  glimpse()

calls_df <- calls \%>\%
  tbl("calls") \%>\%
  collect()
}
}
