% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R, R/etl_transform.R
\name{etl_extract.etl_idswb}
\alias{etl_extract.etl_idswb}
\alias{etl_transform.etl_idswb}
\title{etl_extract}
\usage{
\method{etl_extract}{etl_idswb}(obj, ...)

\method{etl_transform}{etl_idswb}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link[etl]{etl}} object}

\item{...}{arguments passed to methods}
}
\description{
etl_extract

etl_transform
}
\details{
This function downloads International Debt statistics data
in the raw directory under the folder the user created

This function transforms International Debt statistics data for years and months specified.
}
