% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_idswb}
\alias{etl_extract.etl_idswb}
\title{Extract data from the Internet and store it locally in its raw form}
\usage{
\method{etl_extract}{etl_idswb}(obj, ...)
}
\arguments{
\item{obj}{An etl object}

\item{...}{arguments passed to methods}
}
\value{
Populates raw directory with data
}
\description{
Extract data from the Internet and store it locally in its raw form
}
