% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_idswb}
\alias{etl_extract.etl_idswb}
\title{Extracts data and stores it locally in its raw form}
\usage{
\method{etl_extract}{etl_idswb}(obj, ...)
}
\arguments{
\item{obj}{an etl object}

\item{...}{arguments passed to methods}
}
\description{
Downloads data and stores it locally in its raw form
}
